+++ CHANGELOG.rst
@@ -1,6 +1,7 @@
 **2.2.0 - 10/19/23**
 
- - Refactor lookup package and Manager configuration defaults
+ - Refactor Manager configuration defaults
+ - Refactor and simplify LookupTable implementations
  - Enable LookupTable to have source data that is only categorical
  - Throw an error if simulation attempts to use a component that is not an instance of Component